Refactor code for deploy actions - Remove redundant case statements from `slack.ParseAction` - Use DeployAction struct uniformly in all routines - Remove `controllers.helper.go`, shifted DeployAction struct to `handlers.go`
